+++ example/lib/screen/home.dart
@@ -1,5 +1,3 @@
-import 'dart:math';
-
 import 'package:arcane/arcane.dart';
 
 List<String> list = List.generate(5, (index) => "Item $index");
@@ -28,80 +26,39 @@ class _HomeScreenState extends State<HomeScreen> {
         ),
         child: Collection(
           children: [
-            Section(
-                headerText: "The List",
-                child: ArcaneList(
-                    items: list,
-                    itemBuilder: (BuildContext context, int index) {
-                      return Container(
-                        key: ValueKey(list[index]),
-                        decoration: BoxDecoration(
-                            color: HSLColor.fromAHSL(
-                                    0.1,
-                                    Random(list[index].hashCode).nextDouble() *
-                                        360,
-                                    1,
-                                    0.6)
-                                .toColor(),
-                            borderRadius: BorderRadius.circular(8)),
-                        child: ListTile(
-                          onPressed: () {
-                            setState(() {
-                              list.removeAt(index);
-                            });
-                          },
-                          title: Text(list[index]),
-                        ),
-                      );
-                    },
-                    onReorder: (int oldIndex, int newIndex) {
-                      setState(() {
-                        String v = list.removeAt(oldIndex);
-                        list.insert(newIndex, v);
-                      });
-                    },
-                    isSameItem: (a, b) => a == b)),
-            Section(
-                headerText: "The Grid",
-                child: ArcaneGrid(
-                    sliverGridDelegate:
-                        const SliverGridDelegateWithMaxCrossAxisExtent(
-                            childAspectRatio: 2,
-                            mainAxisSpacing: 8,
-                            crossAxisSpacing: 8,
-                            maxCrossAxisExtent: 200),
-                    items: list,
-                    itemBuilder: (BuildContext context, int index) {
-                      return BasicCard(
-                        borderColor: HSLColor.fromAHSL(
-                                0.333,
-                                Random(list[index].hashCode).nextDouble() * 360,
-                                1,
-                                0.6)
-                            .toColor(),
-                        filled: true,
-                        fillColor: HSLColor.fromAHSL(
-                                0.05,
-                                Random(list[index].hashCode).nextDouble() * 360,
-                                1,
-                                0.6)
-                            .toColor(),
-                        onPressed: () {
-                          setState(() {
-                            list.removeAt(index);
-                          });
-                        },
-                        title: Text(list[index]),
-                        key: ValueKey(list[index]),
-                      );
-                    },
-                    onReorder: (int oldIndex, int newIndex) {
-                      setState(() {
-                        String v = list.removeAt(oldIndex);
-                        list.insert(newIndex, v);
-                      });
-                    },
-                    isSameItem: (a, b) => a == b))
+            Card(
+              child: ListTile(
+                title: Text("Title"),
+                subtitle: Text("Subtitle"),
+                leading: Icon(Icons.activity),
+                children: [
+                  ListTile(
+                    title: Text("Title"),
+                    subtitle: Text("Subtitle"),
+                    leading: Icon(Icons.activity),
+                    trailing: Icon(Icons.airplane),
+                  ),
+                  ListTile(
+                    title: Text("Title"),
+                    subtitle: Text("Subtitle"),
+                    leading: Icon(Icons.activity),
+                    trailing: Icon(Icons.airplane),
+                  ),
+                  ListTile(
+                    title: Text("Title"),
+                    subtitle: Text("Subtitle"),
+                    leading: Icon(Icons.activity),
+                    trailing: Icon(Icons.airplane),
+                  ),
+                  ListTile(
+                    title: Text("Title"),
+                    subtitle: Text("Subtitle"),
+                    leading: Icon(Icons.activity),
+                    trailing: Icon(Icons.airplane),
+                  )
+                ],
+              ),
+            )
           ],
         ),
       );